backend/requirements.txt
# FlexFlow Backend - Python 3.10+
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# LiveKit Agents & Gemini
livekit-agents[google]~=1.3
livekit-api>=0.8.0

# Vision (MediaPipe Holistic - in-memory only)
mediapipe>=0.10.0
opencv-python-headless>=4.10.0

# Async
anyio>=4.0.0
